Move the Sparql client around
